CLI: restructure commands
Fixes CODY-2634

- `cody jsonrpc` is now `cody api jsonrpc-stdio`.
- `cody server` is now `cody api jsonrpc-websocket`.
- Files consistently use the `command-NAME.ts` name convention making it
  easier to find them in the fuzzy finder.
