+++ agent/src/cli/command-auth/AuthenticatedAccount.ts
