+++ agent/src/cli/command-auth/command-accounts.ts
