+++ agent/src/cli/command-auth/command-auth.ts
