+++ agent/src/cli/command-auth/command-login.ts
