+++ agent/src/cli/command-auth/command-logout.ts
