+++ agent/src/cli/command-auth/command-whoami.ts
