+++ agent/src/cli/command-auth/messages.ts
