+++ agent/src/cli/command-auth/secrets.ts
