+++ agent/src/cli/command-auth/settings.ts
