+++ agent/src/cli/command-bench/EvaluationDocument.ts
@@ -12,7 +12,7 @@ import { ProtocolTextDocumentWithUri } from '../../../../vscode/src/jsonrpc/Text
 import { AgentTextDocument } from '../../AgentTextDocument'
 
 import type { AutocompleteMatchKind } from './AutocompleteMatcher'
-import { BenchStrategy, type CodyBenchOptions } from './cody-bench'
+import { BenchStrategy, type CodyBenchOptions } from './command-bench'
 import type { EvaluateFileParams } from './evaluateEachFile'
 
 export type EvaluationDocumentParams = Pick<
