+++ agent/src/cli/command-bench/Queries.ts
