+++ agent/src/cli/command-bench/Timer.ts
