+++ agent/src/cli/command-bench/isWindows.ts
