+++ agent/src/cli/command-bench/llm-judge-chat-template.ts
