+++ agent/src/cli/command-bench/matchesGlobPatterns.test.ts
