+++ agent/src/cli/command-bench/matchesGlobPatterns.ts
