+++ agent/src/cli/command-bench/prettyDiagnostic.ts
