+++ agent/src/cli/command-bench/queries/cpp/context.scm
