+++ agent/src/cli/command-bench/queries/csharp/context.scm
