+++ agent/src/cli/command-bench/queries/tsx/context.scm
