+++ agent/src/cli/command-bench/strategy-autocomplete.ts
@@ -3,7 +3,7 @@ import type { RpcMessageHandler } from '../../jsonrpc-alias'
 import { execSync } from 'node:child_process'
 import { type AutocompleteMatchKind, AutocompleteMatcher } from './AutocompleteMatcher'
 import { EvaluationDocument } from './EvaluationDocument'
-import type { CodyBenchOptions } from './cody-bench'
+import type { CodyBenchOptions } from './command-bench'
 import { evaluateEachFile } from './evaluateEachFile'
 import { matchesGlobPatterns } from './matchesGlobPatterns'
 import { triggerAutocomplete } from './triggerAutocomplete'
