+++ agent/src/cli/command-bench/strategy-chat.ts
@@ -5,7 +5,7 @@ import * as vscode from 'vscode'
 import YAML from 'yaml'
 import type { RpcMessageHandler } from '../../jsonrpc-alias'
 import { EvaluationDocument } from './EvaluationDocument'
-import type { CodyBenchOptions } from './cody-bench'
+import type { CodyBenchOptions } from './command-bench'
 import { evaluateEachFile } from './evaluateEachFile'
 import { LlmJudge, type LlmJudgeScore } from './llm-judge'
 import { concisenessPrompt, helpfulnessPrompt } from './llm-judge-chat-template'
