+++ agent/src/cli/command-bench/strategy-fix.ts
@@ -11,7 +11,7 @@ import type { RpcMessageHandler } from '../../jsonrpc-alias'
 import { renderUnifiedDiff } from '../../renderUnifiedDiff'
 import { vscodeRange } from '../../vscode-type-converters'
 import { EvaluationDocument } from './EvaluationDocument'
-import type { CodyBenchOptions } from './cody-bench'
+import type { CodyBenchOptions } from './command-bench'
 import { evaluateEachFile } from './evaluateEachFile'
 import { LlmJudge, type LlmJudgeScore } from './llm-judge'
 import { llmJudgeFixTemplate } from './llm-judge-fix-template'
