+++ agent/src/cli/command-bench/strategy-unit-test.ts
@@ -9,7 +9,7 @@ import { TestClient } from '../../TestClient'
 import { getLanguageForFileName } from '../../language'
 import type { ProtocolDiagnostic } from '../../protocol-alias'
 import { EvaluationDocument } from './EvaluationDocument'
-import type { CodyBenchOptions } from './cody-bench'
+import type { CodyBenchOptions } from './command-bench'
 import { evaluateEachFile } from './evaluateEachFile'
 import { prettyDiagnostic } from './prettyDiagnostic'
 import { runVoidCommand } from './testTypecheck'
