+++ agent/src/cli/command-bench/testParse.ts
